Add Metastore In Memory cache implementation
